Fix : Modify Image Change Logic

+++ posts/serializers.py
@@ -70,10 +70,11 @@ class EcoCarpingSerializer(TaggitSerializer, ModelSerializer):
             elif key == "image3":
                 if not instance.image3 == "" and not instance.image3 == None:
                     s3.delete_file(str(instance.image3))
-            else:
+            elif key == "image4":
                 if not instance.image4 == "" and not instance.image4 == None:
                     s3.delete_file(str(instance.image4))
-
+            else:
+                pass
         return super().update(instance, validated_data)
 
 
@@ -216,10 +217,11 @@ class ShareSerializer(TaggitSerializer, ModelSerializer):
             elif key == "image3":
                 if not instance.image3 == "" and not instance.image3 == None:
                     s3.delete_file(str(instance.image3))
-            else:
+            elif key == "image4":
                 if not instance.image4 == "" and not instance.image4 == None:
                     s3.delete_file(str(instance.image4))
-
+            else:
+                pass
         return super().update(instance, validated_data)
